# swagger.yaml
openapi: 3.0.0
info:
  title: TiketQ Bosbiller API
  description: API for flight searches, bookings, and user management.
  version: 1.0.0
servers:
  - url: /
    description: Local server
paths:
...
  /api/auth:
    post:
      summary: Login user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Login successful
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: User registered
  /api/auth/me:
    get:
      summary: Get current user profile
      description: Returns the profile of the currently logged-in user based on the JWT token.
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
        "401":
          description: Unauthorized
    post:
      summary: Admin login
      tags:
        - Auth
      responses:
        "200":
          description: Admin login successful
  /api/auth/admin-register:
    post:
      summary: Admin registration (Protected)
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Admin registered
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT